Rename service&proposer packages to registartion&validator
