+++ ethereumj-core/src/main/java/org/ethereum/sharding/config/BeaconConfig.java
@@ -43,18 +43,18 @@ import org.ethereum.sharding.processing.db.BeaconStore;
 import org.ethereum.sharding.processing.db.IndexedBeaconStore;
 import org.ethereum.sharding.processing.state.BeaconStateRepository;
 import org.ethereum.sharding.processing.state.StateRepository;
-import org.ethereum.sharding.proposer.BeaconProposer;
-import org.ethereum.sharding.proposer.BeaconProposerImpl;
-import org.ethereum.sharding.proposer.ValidatorService;
-import org.ethereum.sharding.proposer.ValidatorServiceImpl;
-import org.ethereum.sharding.service.MultiValidatorRegistrationService;
-import org.ethereum.sharding.service.ValidatorRepositoryImpl;
-import org.ethereum.sharding.service.ValidatorRegistrationService;
+import org.ethereum.sharding.validator.BeaconProposer;
+import org.ethereum.sharding.validator.BeaconProposerImpl;
+import org.ethereum.sharding.validator.ValidatorService;
+import org.ethereum.sharding.validator.ValidatorServiceImpl;
+import org.ethereum.sharding.registration.MultiValidatorRegistrationService;
+import org.ethereum.sharding.registration.ValidatorRepositoryImpl;
+import org.ethereum.sharding.registration.ValidatorRegistrationService;
 import org.ethereum.sharding.crypto.DepositAuthority;
 import org.ethereum.sharding.contract.DepositContract;
 import org.ethereum.sharding.crypto.UnsecuredDepositAuthority;
 import org.ethereum.sharding.util.Randao;
-import org.ethereum.sharding.service.ValidatorRepository;
+import org.ethereum.sharding.registration.ValidatorRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Condition;
