+++ ethereumj-core/src/main/java/org/ethereum/sharding/manager/ShardingWorldManager.java
@@ -32,11 +32,11 @@ import org.ethereum.listener.EthereumListenerAdapter;
 import org.ethereum.manager.WorldManager;
 import org.ethereum.sharding.config.DepositContractConfig;
 import org.ethereum.sharding.processing.BeaconChain;
-import org.ethereum.sharding.proposer.ValidatorService;
+import org.ethereum.sharding.validator.ValidatorService;
 import org.ethereum.sharding.pubsub.BeaconChainSynced;
 import org.ethereum.sharding.pubsub.Publisher;
 import org.ethereum.sharding.pubsub.ValidatorStateUpdated;
-import org.ethereum.sharding.service.ValidatorRegistrationService;
+import org.ethereum.sharding.registration.ValidatorRegistrationService;
 import org.ethereum.vm.program.invoke.ProgramInvokeFactoryImpl;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
