+++ ethereumj-core/src/main/java/org/ethereum/sharding/registration/ValidatorRepository.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with the ethereumJ library. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.ethereum.sharding.service;
+package org.ethereum.sharding.registration;
 
 import org.ethereum.core.Genesis;
 import org.ethereum.sharding.domain.Validator;
