+++ ethereumj-core/src/main/java/org/ethereum/sharding/registration/ValidatorRepositoryImpl.java
@@ -15,10 +15,9 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with the ethereumJ library. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.ethereum.sharding.service;
+package org.ethereum.sharding.registration;
 
 import org.ethereum.core.Block;
-import org.ethereum.core.Genesis;
 import org.ethereum.core.TransactionInfo;
 import org.ethereum.db.BlockStore;
 import org.ethereum.db.TransactionStore;
