+++ ethereumj-core/src/test/java/org/ethereum/sharding/ShardingTestHelper.java
@@ -41,10 +41,10 @@ import org.ethereum.sharding.contract.DepositContract;
 import org.ethereum.sharding.crypto.DepositAuthority;
 import org.ethereum.sharding.crypto.UnsecuredDepositAuthority;
 import org.ethereum.sharding.pubsub.Publisher;
-import org.ethereum.sharding.service.ValidatorRepository;
-import org.ethereum.sharding.service.ValidatorRepositoryImpl;
-import org.ethereum.sharding.service.ValidatorRegistrationService;
-import org.ethereum.sharding.service.ValidatorRegistrationServiceImpl;
+import org.ethereum.sharding.registration.ValidatorRepository;
+import org.ethereum.sharding.registration.ValidatorRepositoryImpl;
+import org.ethereum.sharding.registration.ValidatorRegistrationService;
+import org.ethereum.sharding.registration.ValidatorRegistrationServiceImpl;
 import org.ethereum.sharding.util.Randao;
 import org.ethereum.util.blockchain.StandaloneBlockchain;
 import org.ethereum.vm.program.ProgramResult;
