+++ ethereumj-core/src/test/java/org/ethereum/sharding/registration/ValidatorRepositoryTest.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with the ethereumJ library. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.ethereum.sharding.service;
+package org.ethereum.sharding.registration;
 
 import org.ethereum.core.Block;
 import org.ethereum.core.Transaction;
